--- .github/workflows/integration_tests.yml
@@ -1,15 +1,0 @@
-name: Integration Tests
-on:
-  pull_request:
-    types: [assigned, opened, synchronize, reopened]
-    paths: ['packages/api/**']
-
-env:
-  CON_STR: '${{ secrets.CON_STR }}'
-jobs:
-  run_tests:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Run Integration Tests
-        run: cd ./packages/api && docker build . --target test --build-arg CON_STR=$CON_STR
