--- packages/api/controllers/articles_controller.go
@@ -1,182 +1,0 @@
-package controllers
-
-import (
-	"context"
-	"hub-backend/configs"
-	"hub-backend/models"
-	"hub-backend/responses"
-	"net/http"
-	"time"
-
-	"github.com/gofiber/fiber/v2"
-	"go.mongodb.org/mongo-driver/bson"
-	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo"
-)
-
-var articlesCollection *mongo.Collection = configs.GetCollection(configs.DB, "articles")
-
-func CreateArticle(c *fiber.Ctx) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-
-	bearer_token := c.Get("BEARER-TOKEN")
-
-	var article models.Article
-	defer cancel()
-
-	if bearer_token != configs.ReturnAuthToken() {
-		return c.Status(http.StatusUnauthorized).JSON(responses.MemberResponse{Status: http.StatusUnauthorized, Message: "error", Data: &fiber.Map{"data": "Unauthorized"}})
-	}
-
-	if err := c.BodyParser(&article); err != nil {
-		return c.Status(http.StatusBadRequest).JSON(responses.MemberResponse{Status: http.StatusBadRequest, Message: "error", Data: &fiber.Map{"data": err.Error()}})
-	}
-
-	if validationErr := validate.Struct(&article); validationErr != nil {
-		return c.Status(http.StatusBadRequest).JSON(responses.MemberResponse{Status: http.StatusBadRequest, Message: "error", Data: &fiber.Map{"data": validationErr.Error()}})
-	}
-
-	// v := reflect.ValueOf(article)
-	// type_of_v := v.Type()
-
-	// // this will only work if all struct fields are string
-	// for i := 0; i < v.NumField(); i++ {
-	// 	empty_body_check := EmptyStringBody(v.Field(i).Interface().(string))
-	// 	if empty_body_check {
-	// 		return isEmptyException(c, type_of_v.Field(i).Name)
-	// 	}
-	// }
-
-	result, err := articlesCollection.InsertOne(ctx, article)
-	if err != nil {
-		return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: "error", Data: &fiber.Map{"data": err.Error()}})
-	}
-
-	return c.Status(http.StatusCreated).JSON(responses.MemberResponse{Status: http.StatusCreated, Message: "success", Data: &fiber.Map{"data": result.InsertedID}})
-
-}
-
-func GetAllArticles(c *fiber.Ctx) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	var article []models.Article
-	defer cancel()
-
-	results, err := articlesCollection.Find(ctx, bson.M{})
-
-	if err != nil {
-		return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: "error", Data: &fiber.Map{"data": err.Error()}})
-	}
-
-	defer results.Close(ctx)
-	for results.Next(ctx) {
-		var singleArticle models.Article
-		if err = results.Decode(&singleArticle); err != nil {
-			return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: "error", Data: &fiber.Map{"data": err.Error()}})
-		}
-
-		article = append(article, singleArticle)
-	}
-
-	return c.Status(http.StatusOK).JSON(
-		responses.MemberResponse{Status: http.StatusOK, Message: "success", Data: &fiber.Map{"data": article}},
-	)
-}
-
-func EditArticle(c *fiber.Ctx) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	key := c.Params("key", "key was not provided")
-	var article models.ArticleUpdate
-
-	bearer_token := c.Get("BEARER-TOKEN")
-
-	defer cancel()
-	if bearer_token != configs.ReturnAuthToken() {
-		return c.Status(http.StatusUnauthorized).JSON(responses.MemberResponse{Status: http.StatusUnauthorized, Message: "error", Data: &fiber.Map{"Reason": "Authentication failed"}})
-	}
-
-	if err := c.BodyParser(&article); err != nil {
-		return c.Status(http.StatusBadRequest).JSON(responses.MemberResponse{Status: http.StatusBadRequest, Message: "Empty Body"})
-	}
-
-	if validationErr := validate.Struct(&article); validationErr != nil {
-		return c.Status(http.StatusBadRequest).JSON(responses.MemberResponse{Status: http.StatusBadRequest, Message: "Body is not compatible"})
-	}
-
-	key_from_hex, _ := primitive.ObjectIDFromHex(key)
-
-	article_map := make(map[string]string)
-
-	if article.Author != "" {
-		article_map["author"] = article.Author
-	}
-	if article.Banner != "" {
-		article_map["banner"] = article.Banner
-	}
-	if article.MediumLink != "" {
-		article_map["mediumLink"] = article.MediumLink
-	}
-	if article.Title != "" {
-		article_map["title"] = article.Title
-	}
-
-	update := bson.M{}
-	for k, v := range article_map {
-		update[k] = v
-	}
-
-	result, err := articlesCollection.UpdateOne(ctx, bson.M{"_id": key_from_hex}, bson.M{"$set": update})
-
-	if err != nil {
-		return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: err.Error()})
-	}
-
-	if result.MatchedCount != 1 {
-		return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: "Document not found"})
-	}
-
-	return c.Status(http.StatusOK).JSON(responses.MemberResponse{Status: http.StatusOK, Message: "User was updated"})
-}
-
-func GetArticle(c *fiber.Ctx) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	key := c.Params("key", "key was not provided")
-
-	var article models.Article
-	defer cancel()
-
-	key_from_hex, _ := primitive.ObjectIDFromHex(key)
-	err := articlesCollection.FindOne(ctx, bson.M{"_id": key_from_hex}).Decode(&article)
-	if err != nil {
-		return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: "error", Data: &fiber.Map{"data": err.Error() + "key: " + key}})
-	}
-
-	return c.Status(http.StatusOK).JSON(responses.MemberResponse{Status: http.StatusOK, Message: "success", Data: &fiber.Map{"data": article}})
-}
-
-func DeleteArticle(c *fiber.Ctx) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	key := c.Params("key", "key was not provided")
-	bearer_token := c.Get("BEARER-TOKEN")
-	defer cancel()
-
-	if bearer_token != configs.ReturnAuthToken() {
-		return c.Status(http.StatusUnauthorized).JSON(responses.MemberResponse{Status: http.StatusUnauthorized, Message: "error", Data: &fiber.Map{"Reason": "Authentication failed"}})
-	}
-
-	key_from_hex, _ := primitive.ObjectIDFromHex(key)
-
-	result, err := articlesCollection.DeleteOne(ctx, bson.M{"_id": key_from_hex})
-	if err != nil {
-		return c.Status(http.StatusInternalServerError).JSON(responses.MemberResponse{Status: http.StatusInternalServerError, Message: "Error", Data: &fiber.Map{"Reason": err.Error(), "Key": key}})
-	}
-
-	if result.DeletedCount < 1 {
-		return c.Status(http.StatusNotFound).JSON(
-			responses.MemberResponse{Status: http.StatusNotFound, Message: "error", Data: &fiber.Map{"data": "Article with specified ID not found!"}},
-		)
-	}
-
-	return c.Status(http.StatusOK).JSON(
-		responses.MemberResponse{Status: http.StatusOK, Message: "success", Data: &fiber.Map{"data": "Article successfully deleted!"}},
-	)
-}
