--- packages/api/routes/hackathon_register_route.go
@@ -1,12 +1,0 @@
-package routes
-
-import (
-	"hub-backend/controllers"
-
-	"github.com/gofiber/fiber/v2"
-)
-
-func HackathonRegisterRoutes(app *fiber.App) {
-	app.Post("/api/hackathon/register", controllers.RegisterTeamMember)
-	app.Get("/api/hackathon/register/available", controllers.IsRegistrationAvailable)
-}
