--- packages/api/routes/hackathon_team_route.go
@@ -1,15 +1,0 @@
-package routes
-
-import (
-	"hub-backend/controllers"
-
-	"github.com/gofiber/fiber/v2"
-)
-
-func HackathonTeamsRoutes(app *fiber.App) {
-	app.Post("/api/hackathon/teams", controllers.CreateHackathonTeam)
-	app.Get("/api/hackathon/teams", controllers.GetHackathonTeams)
-	app.Get("/api/hackathon/teams/:key", controllers.GetHackathonTeam)
-	app.Put("/api/hackathon/teams/:key", controllers.EditHackathonTeams)
-	app.Delete("/api/hackathon/teams/:key", controllers.DeleteHackathonTeams)
-}
