--- packages/api/tests/TESTS_README.md
@@ -1,23 +1,0 @@
-# Test Setup
-## How to run the tests
-```
-# .env
-IS_TEST=true
-```
-
-## .env file in `/tests` (old)
-```
-MONGOURI=...
-IS_TEST=true
-```
-* SYMLINK (AUTHORIZATION TOKEN will be `TEST_TOKEN`)
-```bash
-ln -s ./.env ./tests/.env  # you should be in the root dir
-# add IS_TEST=true
-```
-
-* COPY (AUTHORIZATION TOKEN will still be generated by the cron job)
-```bash
-cp ./.env ./tests/.env  # you should be in the root dir
-# add IS_TEST=true
-```
